Fix: Fixed blob as_sqlite() to use int and not char in output (was assumed text)

+++ senc/utils/sqlite/values.cpp
@@ -323,7 +323,7 @@ namespace senc::utils::sqlite
 		std::stringstream s;
 		s << "x'";
 		for (byte b : get())
-			s << std::hex << b;
+			s << std::hex << static_cast<int>(b);
 		s << "'";
 		return s.str();
 	}
